PPP: Dumps of packets have non-printing (octal) chars sign-extended
Pppd calls print_string() to dump packet contents, using it to print
graphic characters as themselves and converting non-printing characters
to \ooo sequences. Problem is with the latter: the 'char' is converted
to an 'int' by the compiler implicitly, and sign extended. Fix is to
have an intermediate cast to 'unsigned char' so that the promotion to
'int' doesn't involve sign extension.

This fix was submitted upstream already as:

https://github.com/pprindeville/ppp/commit/5e8c3cb256a7e86e3572a82a75d51c6850efdbd

VRVDR-3923